lumistakes/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.4
django-environ==0.12.0
django-rest-framework==0.1.0
django-webpack-loader==3.1.1
djangorestframework==3.16.0
idna==3.10
mysqlclient==2.2.7
oauthlib==3.2.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
tzdata==2024.2
urllib3==2.3.0
